Working on overall compatibility of the glyph set

+++ assistantLib/assistantParts/spacingKerning/kerningManager.py
@@ -1165,7 +1165,7 @@ class KerningManager:
         Print a message and answer the True if the glyph margin was changed by a new value."""
         changed = False
         blm = self.getBasedLeftMargin(g)
-        if abs(lm - blm) >= 1:
+        if None not in (lm, blm) and abs(lm - blm) >= 1:
             print(f'Set based left margin of /{g.name} from {lm:0.2f} to {blm:0.2f} {label}')
             g.angledLeftMargin += lm - blm
             changed = True
@@ -1176,7 +1176,7 @@ class KerningManager:
         Print a message and answer the True if the glyph margin was changed by a new value."""
         changed = False
         brm = self.getBasedRightMargin(g)
-        if abs(rm - brm) >= 1:
+        if None not in (rm, brm)  and abs(rm - brm) >= 1:
             print(f'Set based right margin of /{g.name} from {rm:0.2f} to {brm:0.2f} {label}')
             g.angledRightMargin += rm - brm
             changed = True
